+++ include/pfp_lcp_doc.hpp
@@ -36,6 +36,7 @@ typedef struct
     bool is_start = false;
     bool is_end = false;
     size_t lcp_with_prev_suffix = 0;
+    size_t sa_i = 0;
 } queue_entry_t;
 
 std::ostream& operator<<(std::ostream& os, const queue_entry_t& lcp_queue_entry) {
@@ -44,7 +45,8 @@ std::ostream& operator<<(std::ostream& os, const queue_entry_t& lcp_queue_entry)
                  ", doc num: " << lcp_queue_entry.doc_num << 
                  ", is_start: " << lcp_queue_entry.is_start <<
                  ", is_end: " << lcp_queue_entry.is_end <<
-                 ", lcp_with_prev: " << lcp_queue_entry.lcp_with_prev_suffix << "]";
+                 ", lcp_with_prev: " << lcp_queue_entry.lcp_with_prev_suffix <<
+                 ", pos_of_LFi: " << lcp_queue_entry.sa_i << "]";
 }
 
 class pfp_lcp{
@@ -62,390 +64,19 @@ public:
     size_t NUMCOLSFORTABLE = 0; 
 
     /*
-     * Second alternate approach: avoids usage of predecessor lcp table
-     * since it is time-consuming to continue updating a table that has
-     * a size of num_docs x num_characters
-     */
-    pfp_lcp(size_t random_param, size_t random_param2, size_t random_param3, pf_parsing &pfp_, std::string filename, RefBuilder* ref_build, bool rle_ = true) : 
-                pf(pfp_),
-                min_s(1, pf.n),
-                pos_s(1,0),
-                head(0),
-                num_docs(ref_build->num_docs),
-                ch_doc_counters(256, std::vector<size_t>(ref_build->num_docs, 0)),
-                ch_doc_encountered(256, std::vector<bool>(ref_build->num_docs, false)),
-                predecessor_max_lcp(256, std::vector<size_t>(ref_build->num_docs, ref_build->total_length)),
-                queue_pos_per_tuple(256, std::vector<std::deque<size_t>>(ref_build->num_docs, std::deque<size_t>(0))),
-                rle(rle_)
-                // heads(1, 0)
-    {        
-        // Opening output files
-        std::string outfile = filename + std::string(".lcp");
-        if ((lcp_file = fopen(outfile.c_str(), "w")) == nullptr)
-            error("open() file " + outfile + " failed");
-
-        outfile = filename + std::string(".ssa");
-        if ((ssa_file = fopen(outfile.c_str(), "w")) == nullptr)
-            error("open() file " + outfile + " failed");
-
-        outfile = filename + std::string(".esa");
-        if ((esa_file = fopen(outfile.c_str(), "w")) == nullptr)
-            error("open() file " + outfile + " failed");
-
-        if (rle) {
-            outfile = filename + std::string(".bwt.heads");
-            if ((bwt_file = fopen(outfile.c_str(), "w")) == nullptr)
-                error("open() file " + outfile + " failed");
-
-            outfile = filename + std::string(".bwt.len");
-            if ((bwt_file_len = fopen(outfile.c_str(), "w")) == nullptr)
-                error("open() file " + outfile + " failed");
-        } else {
-            outfile = filename + std::string(".bwt");
-            if ((bwt_file = fopen(outfile.c_str(), "w")) == nullptr)
-                error("open() file " + outfile + " failed");
-        }
-
-        outfile = filename + std::string(".sdap");
-        if ((sdap_file = fopen(outfile.c_str(), "w")) == nullptr) {
-            error("open() file " + outfile + " failed");
-        }
-        if (fwrite(&ref_build->num_docs, sizeof(size_t), 1, sdap_file) != 1)
-            error("SDAP write error: number of documents");
-
-        outfile = filename + std::string(".edap");
-        if ((edap_file = fopen(outfile.c_str(), "w")) == nullptr)
-            error("open() file " + outfile + " failed");
-        if (fwrite(&ref_build->num_docs, sizeof(size_t), 1, edap_file) != 1)
-            error("SDAP write error: number of documents");
-
-        assert(pf.dict.d[pf.dict.saD[0]] == EndOfDict);
-
-        // variables for bwt/lcp/sa construction
-        phrase_suffix_t curr;
-        phrase_suffix_t prev;
-
-        // variables for doc profile construction 
-        uint8_t prev_bwt_ch = 0;
-        size_t curr_run_num = 0;
-        size_t pos = 0;
-        double avg_queue_length = 0.0;
-        size_t total_pos_traversed = 0;
-        std::vector<size_t> curr_da_profile (ref_build->num_docs, 0);
-        std::vector<bool> docs_to_collect (ref_build->num_docs, false);
-
-        // DEBUGGING
-        std::vector<double> times = {0, 0, 0, 0, 0, 0};
-        std::vector<size_t> method_wins = {0, 0, 0};
-        auto start = std::chrono::system_clock::now();
-        auto sec = std::chrono::duration<double>(std::chrono::system_clock::now() - start).count();
-
-        inc(curr);
-        while (curr.i < pf.dict.saD.size())
-        {
-            // Make sure current suffix is a valid proper phrase suffix (at least w characters but not whole phrase)
-            if(is_valid(curr)){
-
-                // Compute the next character of the BWT of T
-                std::vector<phrase_suffix_t> same_suffix(1, curr);
-                phrase_suffix_t next = curr;
-
-                // Go through suffix array of dictionary and store all phrase ids with same suffix
-                while (inc(next) && (pf.dict.lcpD[next.i] >= curr.suffix_length))
-                {
-                    assert(next.suffix_length >= curr.suffix_length);
-                    assert((pf.dict.b_d[next.sn] == 0 && next.suffix_length >= pf.w) || (next.suffix_length != curr.suffix_length));
-                    if (next.suffix_length == curr.suffix_length)
-                    {
-                        same_suffix.push_back(next);
-                    }
-                }
-
-                // Hard case: phrases with different BWT characters precediing them
-                int_t lcp_suffix = compute_lcp_suffix(curr, prev);
-
-                typedef std::pair<int_t *, std::pair<int_t *, uint8_t>> pq_t;
-
-                // using lambda to compare elements.
-                auto cmp = [](const pq_t &lhs, const pq_t &rhs) {
-                    return *lhs.first > *rhs.first;
-                };
-                
-                // Merge a list of occurrences of each phrase in the BWT of the parse
-                std::priority_queue<pq_t, std::vector<pq_t>, decltype(cmp)> pq(cmp);
-                for (auto s: same_suffix)
-                {
-                    size_t begin = pf.pars.select_ilist_s(s.phrase + 1);
-                    size_t end = pf.pars.select_ilist_s(s.phrase + 2);
-                    pq.push({&pf.pars.ilist[begin], {&pf.pars.ilist[end], s.bwt_char}});
-                }
-
-                size_t prev_occ;
-                bool first = true;
-                while (!pq.empty())
-                {
-                    auto curr_occ = pq.top();
-                    pq.pop();
-
-                    if (!first)
-                    {
-                        // Compute the minimum s_lcpP of the the current and previous occurrence of the phrase in BWT_P
-                        lcp_suffix = curr.suffix_length + min_s_lcp_T(*curr_occ.first, prev_occ);
-                    }
-                    first = false;
-
-                    // Update min_s
-                    print_lcp(lcp_suffix, j);
-                    update_ssa(curr, *curr_occ.first);
-                    update_bwt(curr_occ.second.second, 1);
-                    update_esa(curr, *curr_occ.first);
-
-                    ssa = (pf.pos_T[*curr_occ.first] - curr.suffix_length) % (pf.n - pf.w + 1ULL);
-                    esa = (pf.pos_T[*curr_occ.first] - curr.suffix_length) % (pf.n - pf.w + 1ULL);
-
-
-                    /* Start of the DA Profiles code */
-                    
-                    uint8_t curr_bwt_ch = curr_occ.second.second;
-                    size_t lcp_i = lcp_suffix;
-                    size_t sa_i = ssa;
-                    size_t doc_i = ref_build->doc_ends_rank(ssa);
-
-                    // Determine whether current suffix is a run boundary
-                    bool is_start = (pos == 0 || curr_bwt_ch != prev_bwt_ch) ? 1 : 0;
-                    bool is_end = (pos == ref_build->total_length-1); // only special case, common case is below
-
-                    // Handle scenario where the previous suffix was a end of a run (and now we know 
-                    // because we see the next character). So we need to reach into queue.
-                    if (pos != 0 && prev_bwt_ch != curr_bwt_ch) 
-                        lcp_queue.back().is_end = 1; 
-                    
-                    if (is_start) {curr_run_num++;}
-                    size_t pos_of_LF_i = (sa_i > 0) ? (sa_i - 1) : (ref_build->total_length-1);
-                    size_t doc_of_LF_i = ref_build->doc_ends_rank(pos_of_LF_i);
-
-                    // Add the current suffix data to LCP queue 
-                    queue_entry_t curr_entry = {curr_run_num-1, curr_bwt_ch, doc_of_LF_i, is_start, is_end, lcp_i};
-                    lcp_queue.push_back(curr_entry);
-                    ch_doc_counters[curr_bwt_ch][doc_of_LF_i] += 1;
-                    ch_doc_encountered[curr_bwt_ch][doc_of_LF_i] = true;
-
-                    // Prepare variables to use during the lcp queue traversal
-                    size_t min_lcp = lcp_i; // this is lcp with previous suffix
-                    std::fill(docs_to_collect.begin(), docs_to_collect.end(), false);
-                    docs_to_collect[doc_of_LF_i] = true;
-
-                    // Re-initialize doc profiles and max lcp with current document (itself)
-                    std::fill(curr_da_profile.begin(), curr_da_profile.end(), 0);
-                    curr_da_profile[doc_of_LF_i] = ref_build->total_length - pos_of_LF_i;
-
-                    // lambda to check if we haven't found a certain document
-                    auto all = [&](std::vector<bool> docs_collected) {
-                            bool found_all = true;
-                            for (auto elem: docs_collected)
-                                found_all &= elem;
-                            return found_all;
-                    };
-
-                    // Update the vector containing all the lcp values in queue
-                    lcp_vals_in_queue.push_back(lcp_i);
-
-                    // Update the queue_pos lists for the current <ch, doc> pair
-                    queue_pos_per_tuple[curr_bwt_ch][doc_of_LF_i].push_back(pos);
-
-                    // Put together a list of queue positions to traverse. For
-                    // example for <A, 1> with 3 documents, we want to look at all
-                    // <A, 0> and <A, 2> records only until the last previous occurrence
-                    // of <A, 1>.
-                    auto merge_lists = [&](uint8_t curr_row_ch, size_t curr_row_doc) {
-                            std::vector<size_t> queue_pos_for_traversal;
-                            size_t curr_pos_list_size = queue_pos_per_tuple[curr_bwt_ch][curr_row_doc].size();
-
-                            // lower bound position is the previous occurrence of the same bwt ch and document
-                            // pair since we will not have max lcp with any suffixes above that suffix.
-                            size_t lower_bound = (curr_pos_list_size >= 2) ?
-                                                 queue_pos_per_tuple[curr_bwt_ch][curr_row_doc][curr_pos_list_size-2] :
-                                                 0;
-
-                            // Grab the last index for each list, and the documents # with non-zero
-                            // number of entries
-                            std::vector<size_t> docs_with_values;
-                            std::vector<int> iter_pos;
-                            std::vector<bool> docs_seen(num_docs, true);
-                            for (size_t k = 0; k < num_docs; k++) {
-                                iter_pos.push_back(queue_pos_per_tuple[curr_row_ch][k].size()-1);
-                                if (queue_pos_per_tuple[curr_row_ch][k].size()) {
-                                    docs_with_values.push_back(k);
-                                    docs_seen[k] = false;
-                                }
-                            }
-                            
-                            // Merge position lists until we no longer find any positions 
-                            // above our lower bound position
-                            bool still_values_left = true;
-                            while (!all(docs_seen)) {
-                                //still_values_left = false;
-                                for (auto curr_doc: docs_with_values) {
-                                    size_t queue_pos = 0;
-                                    if (iter_pos[curr_doc] >= 0) 
-                                    {
-                                        queue_pos = queue_pos_per_tuple[curr_row_ch][curr_doc][iter_pos[curr_doc]];
-                                        queue_pos_for_traversal.push_back(queue_pos);
-                                        iter_pos[curr_doc]--;
-                                        //still_values_left = true;
-                                        docs_seen[curr_doc] = true;
-                                    }
-                                }
-                            }    
-                            std::sort(queue_pos_for_traversal.begin(), queue_pos_for_traversal.end(), std::greater<size_t>());
-                            return queue_pos_for_traversal;
-                    };
-                   
-                    // Determine which queue positions we need to check and 
-                    // compute the lcp with.
-                    std::vector<size_t> queue_pos_for_traversal = merge_lists(curr_bwt_ch, doc_of_LF_i);
-                    ASSERT(queue_pos_for_traversal.front() == pos,
-                           "Issue with merge_lists occurred");
-                    queue_pos_for_traversal.erase(queue_pos_for_traversal.begin());
-
-
-                    // Go through all the necessary predecessor suffixes to 
-                    // update the profiles
-                    bool passed_same_document = false;
-                    size_t traversal_count = 0;
-                    for (auto queue_pos: queue_pos_for_traversal) {
-                        size_t true_pos = queue_pos - num_records_ejected;
-
-                        ASSERT((lcp_queue[true_pos].bwt_ch == curr_bwt_ch), 
-                               "Issue occurred in doc profile queue traversal");
-                        
-                        // Case 1: Cross same character and document pair
-                        if (lcp_queue[true_pos].doc_num == doc_of_LF_i)
-                            passed_same_document = true;
-                        // Case 2: Cross same character, but a different document
-                        else 
-                        {
-                            size_t min_lcp = *std::min_element(lcp_vals_in_queue.begin()+true_pos+1, 
-                                                               lcp_vals_in_queue.end());
-
-                            // Case 2a: Update the current DA profile since we haven't seen this document yet
-                            if (!docs_to_collect[lcp_queue[true_pos].doc_num]) 
-                            {
-                                curr_da_profile[lcp_queue[true_pos].doc_num] = min_lcp + 1;
-                                docs_to_collect[lcp_queue[true_pos].doc_num] = true;
-                            }
-
-                            // Case 2b: Update the predecessor DA profile if its run boundary
-                            // and we haven't passed the <ch, doc> pair as current suffix
-                            if (!passed_same_document && (lcp_queue[true_pos].is_start or lcp_queue[true_pos].is_end))
-                            {
-                                size_t start_pos = ref_build->num_docs * true_pos;
-                                size_t curr_max_lcp = lcp_queue_profiles[start_pos + doc_of_LF_i];
-                                lcp_queue_profiles[start_pos + doc_of_LF_i] = std::max(curr_max_lcp, min_lcp+1);
-                            }
-                        }
-                        traversal_count++;
-                        // We can safely exit if we have seen all other documents
-                        // and we have passed an occurrence of the same <ch, document> pair
-                        if (all(docs_to_collect) && passed_same_document)
-                            break;
-                    }
-                    avg_queue_length += lcp_queue.size();
-                    total_pos_traversed += traversal_count;
-
-                    //std::cout << lcp_queue.size() << "  " << queue_pos_for_traversal.size() << std::endl;
-                                        
-                    // Add the current profile to the vector (should always be multiple of # of docs)
-                    for (auto elem: curr_da_profile)
-                        lcp_queue_profiles.push_back(elem);
-
-                    assert(lcp_queue_profiles.size() % ref_build->num_docs == 0);
-                    assert(lcp_queue_profiles.size() == (lcp_queue.size() * ref_build->num_docs));
-
-                    // Try to trim the LCP queue and adjust the count matrix ...
-                    // Method #1: non-heuristic by removing entries with multiples 
-                    size_t curr_pos = 0;
-                    size_t records_to_remove_method1 = 0;
-                    while (curr_pos < lcp_queue.size()) {
-                        uint8_t curr_ch = lcp_queue[curr_pos].bwt_ch;
-                        size_t curr_doc = lcp_queue[curr_pos].doc_num;
-                        assert(ch_doc_counters[curr_ch][curr_doc] >= 1);
-
-                        size_t count = 0;
-                        for (size_t i = 0; i < num_docs; i++) {
-                            if (i != curr_doc && ch_doc_counters[curr_ch][i] >= 1)
-                                count++;
-                        }
-
-                        if (ch_doc_counters[curr_ch][curr_doc] >= 1 && count == (num_docs-1))
-                            records_to_remove_method1++;
-                        // TODO: generalize this to take into account characters that only occur once
-                        else if (curr_ch == EndOfDict || (curr_ch != 'A' && curr_ch != 'C'
-                                && curr_ch != 'G' && curr_ch != 'T'))
-                            records_to_remove_method1++;
-                        else
-                            break;
-                        curr_pos++;
-                    }
-
-                    // Makes sure the number of records to remove is not the 
-                    // full queue
-                    size_t records_to_remove = std::min(records_to_remove_method1, lcp_queue.size()-1);
-
-                    // Remove those top n elements, and update counts
-                    update_lcp_queue(records_to_remove, true);
-                    
-                    /* End of the DA Profiles code (except for some update statements below) */
-
-                    // Update prevs
-                    prev_occ = *curr_occ.first;
-                    prev_bwt_ch = curr_bwt_ch;
-
-                    // Update pq
-                    curr_occ.first++;
-                    if (curr_occ.first != curr_occ.second.first)
-                        pq.push(curr_occ);
-
-                    j += 1;
-                    pos += 1;
-                }
-
-                prev = same_suffix.back();
-                curr = next;
-            }
-            else {
-                inc(curr);
-            }
-        }
-
-        // print last BWT char and SA sample
-        print_sa();
-        print_bwt();
-        print_doc_profiles();
-
-        // Close output files
-        fclose(ssa_file);
-        fclose(esa_file);
-        fclose(bwt_file);
-        fclose(lcp_file);
-
-        if (rle)
-            fclose(bwt_file_len);
-    }
-
-
-    /*
      * First alternate approach: Uses a predecessor lcp table to 
      * reduce the number of positions that need to be traversed 
      * in the lcp queue
      */
-    pfp_lcp(bool taxcomp, bool topk, size_t num_cols, pf_parsing &pfp_, std::string filename, RefBuilder* ref_build, bool rle_ = true) : 
+    pfp_lcp(bool use_heuristics, size_t doc_to_extract, bool taxcomp, bool topk, size_t num_cols, pf_parsing &pfp_, 
+            std::string filename, RefBuilder* ref_build, bool rle_ = true) : 
                 pf(pfp_),
                 min_s(1, pf.n),
                 pos_s(1,0),
                 head(0),
                 NUMCOLSFORTABLE(num_cols),
                 num_docs(ref_build->num_docs),
+                doc_to_print_out(doc_to_extract),
                 ch_doc_counters(256, std::vector<size_t>(ref_build->num_docs, 0)),
                 ch_doc_encountered(256, std::vector<bool>(ref_build->num_docs, false)),
                 predecessor_max_lcp(256, std::vector<size_t>(ref_build->num_docs, ref_build->total_length)),
@@ -548,17 +179,18 @@ public:
 
         // DEBUG: output file for storing variables
         // std::ofstream debugging_fd (filename + ".output_data.txt");
-        // leaveout_fd.open(filename + ".leave_one_data.txt");
+        if (doc_to_print_out > 0)
+            leaveout_fd.open(filename + ".extract_dap.txt");
 
         // Create a backup predecessor max lcp table, and re-initialize with max_lcp
         size_t num_blocks_of_32 = num_docs/32;
         num_blocks_of_32++;
 
-        // Since I am using a int array, I want to make sure I don't have overflow, I am using 
-        // a signed int8 since the 512 uint8t load/store were not working.
+        // Initializing the PRED table with length of the text or the 
+        // MAXLCPVALUE which is 2^16 - 1
         uint16_t max_lcp_init = (ref_build->total_length > MAXLCPVALUE) ? MAXLCPVALUE : ref_build->total_length;
-
         uint16_t predecessor_max_lcp_2[256][num_blocks_of_32 * 32]; 
+
         for (size_t i = 0; i < 256; i++)
             for (size_t j = 0; j < num_blocks_of_32 * 32; j++)
                 predecessor_max_lcp_2[i][j] = max_lcp_init;
@@ -652,7 +284,7 @@ public:
                     //debugging_fd << pos << std::setw(20) << (curr_run_num-1) << std::setw(20) << curr_bwt_ch << std::setw(20) << sa_i << std::setw(20) << doc_of_LF_i << std::setw(20) << lcp_i << std::endl;
 
                     // Add the current suffix data to LCP queue 
-                    queue_entry_t curr_entry = {curr_run_num-1, curr_bwt_ch, doc_of_LF_i, is_start, is_end, lcp_i};
+                    queue_entry_t curr_entry = {curr_run_num-1, curr_bwt_ch, doc_of_LF_i, is_start, is_end, lcp_i, pos_of_LF_i};
                     lcp_queue.push_back(curr_entry);
                     ch_doc_counters[curr_bwt_ch][doc_of_LF_i] += 1;
                     ch_doc_encountered[curr_bwt_ch][doc_of_LF_i] = true;
@@ -736,11 +368,11 @@ public:
                     #else                        
                         for (size_t ch_num = 0; ch_num < 256; ch_num++) {
                             for (size_t doc_num = 0; doc_num < num_docs; doc_num++) {
-                                predecessor_max_lcp[ch_num][doc_num] = std::min(predecessor_max_lcp[ch_num][doc_num], std::min(lcp_i, (size_t) MAXLCPVALUE));
+                                predecessor_max_lcp_2[ch_num][doc_num] = std::min(predecessor_max_lcp_2[ch_num][doc_num], (uint16_t) std::min(lcp_i, (size_t) MAXLCPVALUE));
                             }
                         }
                         // Reset the LCP with respect to the current <ch, doc> pair
-                        predecessor_max_lcp[curr_bwt_ch][doc_of_LF_i] = std::min((size_t) MAXLCPVALUE, ref_build->total_length - pos_of_LF_i);
+                        predecessor_max_lcp_2[curr_bwt_ch][doc_of_LF_i] = std::min((size_t) MAXLCPVALUE, ref_build->total_length - pos_of_LF_i);
                     #endif
 
                     /* End of SIMD changes */
@@ -838,7 +470,6 @@ public:
                     if (lcp_queue.size() >= MAXQUEUELENGTH)
                         FATAL_ERROR("queue length during construction grew too large");
                     
-                    
                     // Add the current profile to the vector (should always be multiple of # of docs)
                     for (auto elem: curr_da_profile)
                         lcp_queue_profiles.push_back(elem);
@@ -870,29 +501,32 @@ public:
                     size_t records_to_remove_heuristic = 0, records_to_remove_non_heuristic = 0;
                     bool heuristic_used = false, non_heuristic_used = false;
 
-                    // Method #1 - Heuristic
-                    if (lcp_i <= 8) {
-                        heuristic_used = true;
-                        records_to_remove_heuristic = lcp_queue.size()-1;
-                    // Method #2 - Heuristic 
-                    } else if (lcp_queue.size() % 1000 == 0 && lcp_queue.size() >= 1000) {
-                        size_t non_usual_chars = 0;
-                        auto is_usual_char = [] (auto ch) {return (ch == 'A' || ch == 'C' || ch == 'G' || ch == 'T' || ch == 'U');};
-
-                        for (auto entry: lcp_queue) {
-                            if (!is_usual_char(entry.bwt_ch))
-                                non_usual_chars++;
-                        }
-
-                        if ((non_usual_chars+0.0)/lcp_queue.size() > 0.90) {
-                            records_to_remove_heuristic = lcp_queue.size()-1;
+                    if (use_heuristics) {
+                        // Method #1 - Heuristic
+                        if (lcp_i <= 6) {
                             heuristic_used = true;
-                        }
-                    // Method #3 - Heuristic
-                    } else if (lcp_queue.size() > 2500) {
-                        records_to_remove_heuristic = lcp_queue.size() - 2500;
-                        heuristic_used = true;
-                    } 
+                            records_to_remove_heuristic = lcp_queue.size()-1;
+                        // Method #2 - Heuristic 
+                        } else if (lcp_queue.size() % 1000 == 0 && lcp_queue.size() >= 1000) {
+                            size_t non_usual_chars = 0;
+                            auto is_usual_char = [] (auto ch) {return (ch == 'A' || ch == 'C' || ch == 'G' || ch == 'T' || ch == 'U');};
+
+                            for (auto entry: lcp_queue) {
+                                if (!is_usual_char(entry.bwt_ch))
+                                    non_usual_chars++;
+                            }
+
+                            if ((non_usual_chars+0.0)/lcp_queue.size() > 0.90) {
+                                records_to_remove_heuristic = lcp_queue.size()-1;
+                                heuristic_used = true;
+                            }
+                        // Method #3 - Heuristic
+                        } else if (lcp_queue.size() > 2500) {
+                            records_to_remove_heuristic = lcp_queue.size() - 2500;
+                            heuristic_used = true;
+                        } 
+                    }
+                    
 
                     /* 
                      * If using heuristics did not yield flushing the queue, let's
@@ -942,13 +576,6 @@ public:
                         update_lcp_queue(std::min(records_to_remove_heuristic, lcp_queue.size()-1), true);
                     else if (records_to_remove_non_heuristic > 0)
                         update_lcp_queue(std::min(records_to_remove_non_heuristic, lcp_queue.size()-1), false);
-
-                    // if (curr_run_num == 10) {
-                    //     std::cout << "\n\nFINISHED\npos = " << pos << std::endl;
-                    //     std::exit(1);
-                    // }
-                    
-                    // std::cout << tax_sdap_overflow_ptr << " " << tax_edap_overflow_ptr << std::endl;
                     
                     /* End of the DA Profiles code (except for some update statements below) */
 
@@ -964,7 +591,6 @@ public:
                     j += 1;
                     pos += 1;
                 }
-
                 prev = same_suffix.back();
                 curr = next;
             }
@@ -983,6 +609,7 @@ public:
         fclose(ssa_file); fclose(esa_file);
         fclose(bwt_file);
         fclose(lcp_file);
+        leaveout_fd.close();
 
         if (rle)
             fclose(bwt_file_len);
@@ -996,340 +623,6 @@ public:
             fclose(sdap_file); fclose(edap_file);
         }
     }
-
-    // Original constructor: The main difference is that it will iterate through 
-    // the ALL the queue positions to compute max LCP with predecessor and successor
-    // while the new constructor, will compute predecessor using a table and iterate
-    // through queue to find the max lcp with successor rows.
-    pfp_lcp(pf_parsing &pfp_, std::string filename, RefBuilder* ref_build, bool rle_ = true) : 
-                pf(pfp_),
-                min_s(1, pf.n),
-                pos_s(1,0),
-                head(0),
-                num_docs(ref_build->num_docs),
-                ch_doc_counters(256, std::vector<size_t>(ref_build->num_docs, 0)),
-                rle(rle_)
-                // heads(1, 0)
-    {
-        // Opening output files
-        std::string outfile = filename + std::string(".lcp");
-        if ((lcp_file = fopen(outfile.c_str(), "w")) == nullptr)
-            error("open() file " + outfile + " failed");
-
-        outfile = filename + std::string(".ssa");
-        if ((ssa_file = fopen(outfile.c_str(), "w")) == nullptr)
-            error("open() file " + outfile + " failed");
-
-        outfile = filename + std::string(".esa");
-        if ((esa_file = fopen(outfile.c_str(), "w")) == nullptr)
-            error("open() file " + outfile + " failed");
-
-        if (rle) {
-            outfile = filename + std::string(".bwt.heads");
-            if ((bwt_file = fopen(outfile.c_str(), "w")) == nullptr)
-                error("open() file " + outfile + " failed");
-
-            outfile = filename + std::string(".bwt.len");
-            if ((bwt_file_len = fopen(outfile.c_str(), "w")) == nullptr)
-                error("open() file " + outfile + " failed");
-        } else {
-            outfile = filename + std::string(".bwt");
-            if ((bwt_file = fopen(outfile.c_str(), "w")) == nullptr)
-                error("open() file " + outfile + " failed");
-        }
-
-        outfile = filename + std::string(".sdap");
-        if ((sdap_file = fopen(outfile.c_str(), "w")) == nullptr) {
-            error("open() file " + outfile + " failed");
-        }
-        if (fwrite(&ref_build->num_docs, sizeof(size_t), 1, sdap_file) != 1)
-            error("SDAP write error: number of documents");
-
-        outfile = filename + std::string(".edap");
-        if ((edap_file = fopen(outfile.c_str(), "w")) == nullptr)
-            error("open() file " + outfile + " failed");
-        if (fwrite(&ref_build->num_docs, sizeof(size_t), 1, edap_file) != 1)
-            error("SDAP write error: number of documents");
-
-        assert(pf.dict.d[pf.dict.saD[0]] == EndOfDict);
-
-        // variables for bwt/lcp/sa construction
-        phrase_suffix_t curr;
-        phrase_suffix_t prev;
-
-        // variables for doc profile construction 
-        uint8_t prev_bwt_ch = 0;
-        size_t curr_run_num = 0;
-        size_t pos = 0;
-        std::vector<size_t> curr_da_profile (ref_build->num_docs, 0);
-        std::vector<bool> docs_to_collect (ref_build->num_docs, false);
-
-        // reserve some room for the lcp queue to avoid reallocation/copying
-        //lcp_queue.reserve(100);
-        //lcp_queue_profiles.reserve(100 * ref_build->num_docs);
-
-        inc(curr);
-        while (curr.i < pf.dict.saD.size())
-        {
-            // Make sure current suffix is a valid proper phrase suffix (at least w characters but not whole phrase)
-            if(is_valid(curr)){
-
-                // Compute the next character of the BWT of T
-                std::vector<phrase_suffix_t> same_suffix(1, curr);
-                phrase_suffix_t next = curr;
-
-                // Go through suffix array of dictionary and store all phrase ids with same suffix
-                while (inc(next) && (pf.dict.lcpD[next.i] >= curr.suffix_length))
-                {
-                    assert(next.suffix_length >= curr.suffix_length);
-                    assert((pf.dict.b_d[next.sn] == 0 && next.suffix_length >= pf.w) || (next.suffix_length != curr.suffix_length));
-                    if (next.suffix_length == curr.suffix_length)
-                    {
-                        same_suffix.push_back(next);
-                    }
-                }
-
-                // Hard case: phrases with different BWT characters precediing them
-                int_t lcp_suffix = compute_lcp_suffix(curr, prev);
-
-                typedef std::pair<int_t *, std::pair<int_t *, uint8_t>> pq_t;
-
-                // using lambda to compare elements.
-                auto cmp = [](const pq_t &lhs, const pq_t &rhs) {
-                    return *lhs.first > *rhs.first;
-                };
-                
-                // Merge a list of occurrences of each phrase in the BWT of the parse
-                std::priority_queue<pq_t, std::vector<pq_t>, decltype(cmp)> pq(cmp);
-                for (auto s: same_suffix)
-                {
-                    size_t begin = pf.pars.select_ilist_s(s.phrase + 1);
-                    size_t end = pf.pars.select_ilist_s(s.phrase + 2);
-                    pq.push({&pf.pars.ilist[begin], {&pf.pars.ilist[end], s.bwt_char}});
-                }
-
-                size_t prev_occ;
-                bool first = true;
-                while (!pq.empty())
-                {
-                    auto curr_occ = pq.top();
-                    pq.pop();
-
-                    if (!first)
-                    {
-                        // Compute the minimum s_lcpP of the the current and previous occurrence of the phrase in BWT_P
-                        lcp_suffix = curr.suffix_length + min_s_lcp_T(*curr_occ.first, prev_occ);
-                    }
-                    first = false;
-
-                    // Update min_s
-                    print_lcp(lcp_suffix, j);
-                    update_ssa(curr, *curr_occ.first);
-                    update_bwt(curr_occ.second.second, 1);
-                    update_esa(curr, *curr_occ.first);
-
-                    ssa = (pf.pos_T[*curr_occ.first] - curr.suffix_length) % (pf.n - pf.w + 1ULL);
-                    esa = (pf.pos_T[*curr_occ.first] - curr.suffix_length) % (pf.n - pf.w + 1ULL);
-
-
-                    /* Start of the DA Profiles code */
-
-                    uint8_t curr_bwt_ch = curr_occ.second.second;
-                    size_t lcp_i = lcp_suffix;
-                    size_t sa_i = ssa;
-                    size_t doc_i = ref_build->doc_ends_rank(ssa);
-
-                    // Determine whether current suffix is a run boundary
-                    bool is_start = (pos == 0 || curr_bwt_ch != prev_bwt_ch) ? 1 : 0;
-                    bool is_end = (pos == ref_build->total_length-1); // only special case, common case is below
-
-                    // Handle scenario where the previous suffix was a end of a run (and now we know 
-                    // because we see the next character). So we need to reach into queue.
-                    if (pos != 0 && prev_bwt_ch != curr_bwt_ch) 
-                        lcp_queue.back().is_end = 1; 
-                    
-                    if (is_start) {curr_run_num++;}
-                    size_t pos_of_LF_i = (sa_i > 0) ? (sa_i - 1) : (ref_build->total_length-1);
-                    size_t doc_of_LF_i = ref_build->doc_ends_rank(pos_of_LF_i);
-
-                    // Add the current suffix data to LCP queue 
-                    queue_entry_t curr_entry = {curr_run_num-1, curr_bwt_ch, doc_of_LF_i, is_start, is_end, lcp_i};
-                    lcp_queue.push_back(curr_entry);
-                    ch_doc_counters[curr_bwt_ch][doc_of_LF_i] += 1;
-
-                    // Prepare variables to use during the lcp queue traversal
-                    size_t min_lcp = lcp_i; // this is lcp with previous suffix
-                    bool passed_same_document = false;
-                    std::fill(docs_to_collect.begin(), docs_to_collect.end(), false);
-                    docs_to_collect[doc_of_LF_i] = true;
-
-                    // Re-initialize doc profiles and max lcp with current document (itself)
-                    std::fill(curr_da_profile.begin(), curr_da_profile.end(), 0);
-                    curr_da_profile[doc_of_LF_i] = ref_build->total_length - pos_of_LF_i;
-
-                    // lambda to check if we haven't found a certain document
-                    auto all = [&](std::vector<bool> docs_collected) {
-                            bool found_all = true;
-                            for (auto elem: docs_collected)
-                                found_all &= elem;
-                            return found_all;
-                    };
-
-                    // Update the vector containing all the lcp values in queue
-                    lcp_vals_in_queue.push_back(lcp_i);
-
-                    // get index before current suffix, use int for signedness
-                    int queue_pos = lcp_queue.size() - 2; 
-                    while (queue_pos >= 0 && (!all(docs_to_collect) || !passed_same_document)) 
-                    {  
-                        // Case 1: Cross same character and document pair
-                        if (lcp_queue[queue_pos].bwt_ch == curr_bwt_ch && 
-                            lcp_queue[queue_pos].doc_num == doc_of_LF_i) 
-                        {
-                                passed_same_document = true;
-                        } 
-                        // Case 2: Cross same character, but a different document
-                        else if (lcp_queue[queue_pos].bwt_ch == curr_bwt_ch &&
-                                 lcp_queue[queue_pos].doc_num != doc_of_LF_i) 
-                        {
-                            // Case 2a: Update the current DA profile we haven't seen this document yet
-                            if (!docs_to_collect[lcp_queue[queue_pos].doc_num]) 
-                            {
-                                curr_da_profile[lcp_queue[queue_pos].doc_num] = min_lcp + 1;
-                                docs_to_collect[lcp_queue[queue_pos].doc_num] = true;
-                            }
-
-                            // Case 2b: Update the predecessor DA profile if its run boundary
-                            // and we haven't passed the <ch, doc> pair as current suffix
-                            if (!passed_same_document && (lcp_queue[queue_pos].is_start or lcp_queue[queue_pos].is_end))
-                            {
-                                size_t start_pos = ref_build->num_docs * queue_pos;
-                                size_t curr_max_lcp = lcp_queue_profiles[start_pos + doc_of_LF_i];
-                                lcp_queue_profiles[start_pos + doc_of_LF_i] = std::max(curr_max_lcp, min_lcp+1);
-                            }
-                        }
-                        
-                        min_lcp = std::min(min_lcp, lcp_queue[queue_pos].lcp_with_prev_suffix);
-                        queue_pos--;
-                    }
-
-                    // Add the current profile to the vector (should always be multiple of # of docs)
-                    for (auto elem: curr_da_profile)
-                        lcp_queue_profiles.push_back(elem);
-
-                    assert(lcp_queue_profiles.size() % ref_build->num_docs == 0);
-                    assert(lcp_queue_profiles.size() == (lcp_queue.size() * ref_build->num_docs));
-
-                    // Try to trim the LCP queue and adjust the count matrix ...
-                    // Method #1: non-heuristic by removing entries with multiples 
-                    size_t curr_pos = 0;
-                    size_t records_to_remove = 0;
-                    
-                    while (curr_pos < lcp_queue.size()) {
-                        uint8_t curr_ch = lcp_queue[curr_pos].bwt_ch;
-                        size_t curr_doc = lcp_queue[curr_pos].doc_num;
-                        assert(ch_doc_counters[curr_ch][curr_doc] >= 1);
-
-                        size_t count = 0;
-                        for (size_t i = 0; i < num_docs; i++) {
-                            if (i != curr_doc && ch_doc_counters[curr_ch][i] >= 1)
-                                count++;
-                        }
-
-                        // Means we cannot remove it from LCP queue
-                        if ((ch_doc_counters[curr_ch][curr_doc] > 1 && (count == num_docs - 1))  || (curr_ch == EndOfDict))
-                            records_to_remove++;
-                        else    
-                            break;
-                        curr_pos++;
-                    }
-
-                    // Method #2: heuristic-based, remove records before a 
-                    // a small LCP value (7) since they are insignificant
-                    /*
-                    size_t records_to_remove_for_lcp = 0;
-                    curr_pos = 0;
-                    while (records_to_remove_for_lcp < lcp_queue.size()) {
-                        if (lcp_queue[records_to_remove_for_lcp++].lcp_with_prev_suffix <= 7)
-                            break;
-                    }
-                    records_to_remove_for_lcp = (records_to_remove_for_lcp == lcp_queue.size()) ? (0) : records_to_remove_for_lcp;
-                    */
-
-                    // Method #3: non-heuristic, remove all the records starting
-                    // from the the front of the queue where the smallest doc array
-                    // profile value is larger than the current smallest lcp value in 
-                    // queue, meaning that no successors rows will have the max lcp with
-                    // these records.
-                    /*
-                    size_t records_to_remove_method2 = 0;
-                    curr_pos = 0;
-                    while (curr_pos < lcp_queue.size()) {
-                        // Get the smallest LCP in the current profile we are looking at
-                        size_t start_pos = ref_build->num_docs * curr_pos;
-                        size_t curr_record_min_lcp = ref_build->total_length;
-                        for (size_t i = 0; i < num_docs; i++) 
-                            curr_record_min_lcp = std::min(lcp_queue_profiles[start_pos+i], curr_record_min_lcp);
-                        
-                        // Smallest LCP from current record until end of queue
-                        size_t curr_min = *std::min_element(lcp_vals_in_queue.begin()+curr_pos+1, lcp_vals_in_queue.end());
-
-                        // If the smallest value less than or equal to curr mininimum,
-                        // then we cannot write it out.
-                        if (curr_record_min_lcp <= curr_min && lcp_queue[curr_pos].bwt_ch != EndOfDict)
-                            break;
-                        else 
-                            records_to_remove_method2++;
-                        curr_pos++; 
-                    } */ 
-
-
-                    // Take the maximum value from the two methods above, BUT
-                    // we cannot reduce the queue to empty because we need to 
-                    // make sure to update records that are the end of runs
-                    //records_to_remove = std::max(records_to_remove, records_to_remove_method2);
-                    records_to_remove = std::min(records_to_remove, lcp_queue.size()-1);
-
-                    // Remove those top n elements, and update counts
-                    update_lcp_queue(records_to_remove, true);
-                    
-                    /* End of the DA Profiles code (except for some update statements below) */
-
-                    // Update prevs
-                    prev_occ = *curr_occ.first;
-                    prev_bwt_ch = curr_bwt_ch;
-
-                    // Update pq
-                    curr_occ.first++;
-                    if (curr_occ.first != curr_occ.second.first)
-                        pq.push(curr_occ);
-
-                    j += 1;
-                    pos += 1;
-                }
-
-                prev = same_suffix.back();
-                curr = next;
-            }
-            else {
-                inc(curr);
-            }
-        }
-        // print last BWT char and SA sample
-        print_sa();
-        print_bwt();
-        print_doc_profiles();
-
-        // Close output files
-        fclose(ssa_file);
-        fclose(esa_file);
-        fclose(bwt_file);
-        fclose(lcp_file);
-
-        if (rle)
-            fclose(bwt_file_len);
-    }
-
 
 
 private:
@@ -1377,6 +670,7 @@ private:
     bool use_taxcomp = false;
     bool use_topk = false;
     size_t tax_sdap_overflow_ptr = 0, tax_edap_overflow_ptr = 0;
+    size_t doc_to_print_out = 0;
 
     FILE *sdap_file; // start of document array profiles
     FILE *edap_file; // end of document array profiles
@@ -1420,6 +714,12 @@ private:
                 if (is_end && fwrite(&curr_ch, 1, 1, edap_file) != 1)
                     FATAL_ERROR("issue occurred while writing to *.edap file");
 
+                // Added for -e (extraction) option
+                size_t sa_pos = lcp_queue[i].sa_i;
+                size_t doc_num = lcp_queue[i].doc_num;
+                if (doc_to_print_out > 0 && doc_num+1 == doc_to_print_out)
+                    leaveout_fd << sa_pos << " "; 
+
                 for (size_t j = 0; j < num_docs; j++) {
                     size_t prof_val = std::min(lcp_queue_profiles.front(), (size_t) MAXLCPVALUE);
                     lcp_queue_profiles.pop_front();
@@ -1428,7 +728,14 @@ private:
                         error("SA write error 1");
                     if (is_end && fwrite(&prof_val, DOCWIDTH, 1, edap_file) != 1)
                         error("SA write error 1");
+
+                    // Added for -e (extraction) option
+                    if (doc_to_print_out > 0 && doc_num+1 == doc_to_print_out) 
+                        leaveout_fd << prof_val << " ";
                 }
+                // Added for -e (extraction) option
+                if (doc_to_print_out > 0 && doc_num+1 == doc_to_print_out)
+                    leaveout_fd << "\n";
             } else if (use_taxcomp) {
                 size_t curr_val = 0, prev_max = std::min(lcp_queue_profiles.front(), (size_t) MAXLCPVALUE);
                 left_increases.push_back(0);
@@ -1539,6 +846,12 @@ private:
             bool is_start = lcp_queue[0].is_start;
             bool is_end = lcp_queue[0].is_end;
 
+            // Added for the -e (extraction) option
+            size_t sa_pos = lcp_queue[0].sa_i;
+            size_t doc_num = lcp_queue[0].doc_num;
+            if (doc_to_print_out > 0 && doc_num == doc_to_print_out-1)
+                leaveout_fd << sa_pos << " "; 
+
             // Update <ch, doc> count matrix
             if (update_table)
                 ch_doc_counters[curr_ch][curr_doc] -= 1;
@@ -1572,7 +885,14 @@ private:
                         error("SA write error 1");
                     if (is_end && fwrite(&prof_val, DOCWIDTH, 1, edap_file) != 1)
                         error("SA write error 1");
+                    
+                    // Added for the -e (extraction) option
+                    if (doc_to_print_out > 0 && doc_num == doc_to_print_out-1) 
+                        leaveout_fd << prof_val << " ";
                 }
+                // Added for the -e (extraction) option
+                if (doc_to_print_out > 0 && doc_num == doc_to_print_out-1)
+                    leaveout_fd << "\n";
             } else if (use_taxcomp) {
                 size_t curr_val = 0, prev_max = std::min(lcp_queue_profiles.front(), (size_t) MAXLCPVALUE);
                 left_increases.push_back(0);